refactor : jwt 관련 코드 리팩토링
