+++ next.config.js
@@ -1,6 +1,12 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  images: { domains: ['kr.object.ncloudstorage.com'] },
+  images: {
+    domains: [
+      'kr.object.ncloudstorage.com',
+      'blog.kakaocdn.net',
+      'www.lottehotel.com',
+    ],
+  },
 }
 
 module.exports = nextConfig
