+++ src/components/hotel/register/HotelType.js
@@ -5,7 +5,7 @@ import { Button } from '@nextui-org/react'
 import { FaArrowRight } from 'react-icons/fa'
 import { useRouter } from 'next/navigation'
 import { useRecoilState } from 'recoil'
-import { hotelTypeState } from '@/util/hotelState'
+import { hotelTypeState } from '@/store/hotelState'
 import { getIconForType } from '@/constants/hotel'
 import { hotelTypes } from '@/constants/constants'
 import { toast } from 'react-toastify'
