--- src/util/userState.js
@@ -1,6 +1,0 @@
-import { atom } from 'recoil'
-
-export const isLoggedInState = atom({
-  key: 'isLoggedIn',
-  default: false,
-})
